Codigos de error añadidos

+++ src/controllers/tarjetaRecarga.controller.js
@@ -1,33 +1,48 @@
 import {TarjetaRecarga} from '../models/TarjetaRecarga.js';
-
+//import {Tienda} from '../models/' IMPORTAR TIENDA
 export const createTarjeta = async (req, res) =>{
     try {
-        const {idTarjeta,idTienda,monto} = req.body;
+        const {idTienda,monto} = req.body;
+        if(idTienda>0 && monto>0){
+            //PONER UNA VALIDACION SI EXISTE LA TIENDA
+            //await Tienda.findByPk(idTienda) != null
+            const newTarjeta = await TarjetaRecarga.create({
+                idTienda,
+                monto
+            });
+            res.status(201).json(newTarjeta)
+        }else{
+            res.status(400).json({message: "Datos invalidos"});
+        }
 
-        const newTarjeta = await TarjetaRecarga.create({
-            idTienda,
-            monto
-        });
-
-    res.json(newTarjeta)
     } catch (error) {
+        res.json({error:`${error}`})
         console.log(error)
     }
 }
 
 export const getTarjeta = async(req, res) =>{
     const {idTarjeta} = req.params;
-
-    try {
-        const tarjeta = await TarjetaRecarga.findOne({
-            where: {
-                idTarjeta,
+    if(idTarjeta>0){
+        try {
+            const tarjeta = await TarjetaRecarga.findOne({
+                where: {
+                    idTarjeta,
+                }
+            });
+            if(tarjeta != null){
+                res.status(200).json(tarjeta)
+            }else{
+                res.status(204).json()
             }
-        });
-        res.json(tarjeta)
-    } catch (error) {
-        console.log(error);
-    }    
+        } catch (error) {
+            res.json({error:`${error}`})
+            console.log(error)
+        }    
+
+    }else{
+        res.status(400).json({message: "Datos invalidos"});
+    }
 }
 
 export const updateTarjeta = async(req,res) =>{
@@ -35,15 +50,23 @@ export const updateTarjeta = async(req,res) =>{
         const {idTarjeta} = req.params;
         const {idTienda, monto} = req.body;
 
-        const tarjeta = await TarjetaRecarga.findByPk(idTarjeta);
-        tarjeta.idTienda = idTienda;
-        tarjeta.monto = monto;
-
-        await tarjeta.save();
-        res.json(tarjeta);
-
+        if(idTarjeta>0 && idTarjeta>0){
+            const tarjeta = await TarjetaRecarga.findByPk(idTarjeta);
+            if(tarjeta!=null){
+                tarjeta.idTienda = idTienda;
+                tarjeta.monto = monto;
+        
+                await tarjeta.save();
+                res.status(201).json(tarjeta);
+            }else{
+                res.status(204).json();
+            }
+        }else{
+            res.status(400).json({message: "Datos invalidos"});
+        }
     } catch (error) {
-        console.log(error);
+        res.json({error:`${error}`})
+        console.log(error)
     }
 }
 
@@ -52,14 +75,23 @@ export const updateMonto = async(req,res) =>{
         const {idTarjeta} = req.params;
         const {monto} = req.body;
 
-        const tarjeta = await TarjetaRecarga.findByPk(idTarjeta);
-        tarjeta.monto = monto;
-
-        await tarjeta.save();
-        res.json(tarjeta);
+        if(idTarjeta>0 && monto>0){
+            const tarjeta = await TarjetaRecarga.findByPk(idTarjeta);
+            if(tarjeta!=null){
+                tarjeta.monto += monto;
+        
+                await tarjeta.save();
+                res.status(201).json(tarjeta);
+            }else{
+                res.status(204).json();
+            }
+        }else{
+            res.status(400).json({message: "Datos invalidos"});
+        }
 
     } catch (error) {
-        console.log(error);
+        res.json({error:`${error}`})
+        console.log(error)
     }
 }
 
@@ -67,14 +99,23 @@ export const deleteTarjeta = async(req,res) => {
     try {
         const {idTarjeta} = req.params;
 
-        await TarjetaRecarga.destroy({
-            where: {
-                idTarjeta,
+        if(idTarjeta > 0){
+            if(await TarjetaRecarga.findByPk(idTarjeta) != null){
+                await TarjetaRecarga.destroy({
+                    where: {
+                        idTarjeta,
+                    }
+                });
+                res.status(200).json();
+            }else{
+                res.status(204).json();
             }
-        });
+        }else{
+            res.status(400).json({message: "Datos invalidos"});
+        }
 
-        return res.sendStatus(204);
     } catch (error) {
+        res.json({error:`${error}`})
         console.log(error)
     }
 }
